# docs/swagger.yaml
basePath: /
definitions:
  controllers.ArticleListResp:
    properties:
      collection_count:
        type: integer
      commentcount:
        type: integer
      id:
        type: integer
      likes:
        type: integer
      preview:
        type: string
      repost_count:
        type: integer
      title:
        type: string
      username:
        type: string
    type: object
  controllers.ArticleResp:
    properties:
      created_at:
        type: string
      id:
        type: integer
      likes:
        type: integer
      preview:
        type: string
      title:
        type: string
      username:
        type: string
    type: object
...
  controllers.CreateArticleDTO:
    properties:
      content:
        type: string
      preview:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - content
    - preview
    - title
    type: object
...
  controllers.MyArticleItem:
    properties:
      collection_count:
        type: integer
      comment_count:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      likes:
        type: integer
      preview:
        type: string
      repost_count:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  controllers.UpdateArticleDTO:
    properties:
      content:
        type: string
      preview:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: Go-Web综合性应用API接口文档
  title: Go-Web项目 API
  version: 0.0.3
paths:
...
  /articles:
    get:
      parameters:
      - description: 关键字（匹配文件名，模糊）
        in: query
        name: title
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页的条数（默认10，最大100）
        in: query
        name: page_size
        type: integer
      - description: 排序：共8种组合，两种排序方式-上传日期 created_desc（默认）/created_asc/likes_desc/likes_asc/comments_desc/comments_asc/reposts_desc/reposts_asc/collections_desc/collections_asc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.ArticleListResp'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取全部文章
      tags:
      - Articles
...
  /articles/{id}:
    delete:
      consumes:
      - application/json
      description: 根据文章ID永久删除当前用户拥有的文章。该操作不可恢复。
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 文章ID无效
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权限访问" // 可选，但建议用 404 隐藏存在性
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 文章不存在或无权限
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 永久删除当前用户的文章
      tags:
      - Articles
...
  /articles/me:
    get:
      parameters:
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页条数（默认10，最大50）
        in: query
        name: page_size
        type: integer
      - description: 排序：created_desc(默认)/created_asc/likes_desc/updated_desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.MyArticleItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取当前用户的文章列表（管理用）
      tags:
      - Articles
...
  /create_articles:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建文章参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateArticleDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.ArticleResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 创建文章
      tags:
      - Articles
...
  /update_articles/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新文章参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateArticleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ArticleResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 更新文章数据
      tags:
      - Articles
...
swagger: "2.0"